added line to the getRecord method
so when adding a new person to the list they would get the proper task record count

+++ resources/views/task/assignTask.blade.php
@@ -10,14 +10,14 @@
   
   @if (!auth()->guest())                   
   
-  @if(auth()->user()->role->id === 5)  
+  @if(auth()->user()->employee->role->id === 2)  
   <h4 class="text-uppercase text-dark title py-3">Tareas Programadas</h4>
   @endif
   @endif
-
+  
   @if (!auth()->guest())                   
   
-  @if(auth()->user()->role->id === 1) 
+  @if(auth()->user()->employee->role->id === 1) 
   <div class="title py-3">
     <h4 class="text-uppercase text-dark">Asignación de tareas</h4>
   </div>
@@ -40,57 +40,60 @@
   </div>
   
   @endif
-    @include('result_message')
+  <br>
+  @include('result_message')
+  
   @if (!auth()->guest())                   
   
-  @if(auth()->user()->role->id === 1) 
+  @if(auth()->user()->employee->role->id === 1) 
   
   {{-- Form to asign a task --}}
   <form action="/assigntask" method="post" class="py-3">
     
     {{-- laravel security measure --}}
     @csrf
-     
-      <div class="bg-whitesmoke p-3 btn col-12">
-        <div class="form-row justify-content-center ">
-          {{-- Task option selectors --}}
-          <div class="col-md-4 col-sm-6 col-xs-12">
-            <label for="date">Fecha</label>
-            <input class="form-control " type="date" name="date" id="date" placeholder="Fecha 2015-01-31" required="required">
-          </div>
-          <div class="col-md-3 col-sm-6 col-xs-12">
-            <label for="task">Tarea</label>
-            <select class="form-control"  id="task_assign_task" name="task" required>
-              <option value="">- Tarea -</option>
-              
-              {{-- [landing] Task list extracted from the database --}}
-              @foreach ($tasks as $key => $task)
-              <option value="{{ $task->id }}">{{ $task->name }} </option>
-              @endforeach
-              
-            </select>
-          </div>
-          
-          {{-- Position list based on task selected previously  [upon request][task_position.js] --}}
-          <div class="col-md-3 col-sm-4 col-xs-12">
-            <label for="position">Empleo</label>
-            <select class="form-control " id="position_assign_task" name="position" required>
-              <option value="">- Empleo -</option>
-            </select>
-          </div>
-          
-          {{-- how many people will do the task --}}
-          <div class="col-md-2 col-sm-4 col-xs-12">
-            <label for="quantity">PAX.</label>
-            <select class="form-control " name="quantity" id="quantity" required>
-              <option value="1">1</option>
-              <option value="2">2</option>
-              <option value="3">3</option>
-              <option value="4">4</option>
-              <option value="5">5</option>
-            </select> 
-          </div>
-        </div>  
+    
+    <div class="bg-whitesmoke p-3 btn col-12">
+      <div class="form-row justify-content-center ">
+        {{-- Task option selectors --}}
+        <div class="col-md-4 col-sm-6 col-xs-12">
+          <label for="date">Fecha</label>
+          <input class="form-control " type="date" name="date" id="date" placeholder="Fecha 2015-01-31" required="required">
+        </div>
+        <div class="col-md-3 col-sm-6 col-xs-12">
+          <label for="task">Tarea</label>
+          <select class="form-control"  id="task_assign_task" name="task" required>
+            <option value="">- Tarea -</option>
+            
+            {{-- [landing] Task list extracted from the database --}}
+            @foreach ($tasks as $key => $task)
+            <option value="{{ $task->id }}">{{ $task->name }} </option>
+            @endforeach
+            
+          </select>
+        </div>
+        
+        {{-- Position list based on task selected previously  [upon request][task_position.js] --}}
+        <div class="col-md-3 col-sm-4 col-xs-12">
+          <label for="position">Empleo</label>
+          <select class="form-control " id="position_assign_task" name="position" required>
+            <option value="">- Empleo -</option>
+          </select>
+        </div>
+        
+        {{-- how many people will do the task --}}
+        <div class="col-md-2 col-sm-4 col-xs-12">
+          <label for="quantity">PAX.</label>
+          <select class="form-control " name="quantity" id="quantity" required>
+            <option value="1">1</option>
+            <option value="2">2</option>
+            <option value="3">3</option>
+            <option value="4">4</option>
+            <option value="5">5</option>
+            <option value="10">10</option>
+          </select> 
+        </div>
+      </div>  
       
     </div>
     {{-- submit button --}}
@@ -119,8 +122,12 @@
           <td>{{$task['task']}}</td>
           <td>{{ date('d M Y', strtotime($task['date'])) }}</td>
           <td class="text-center">
-            <a class="delete-link" href="/assignTask/delete/{{$task['id']}}"> Borrar </a>
-
+            @if (!auth()->guest())                   
+            
+              @if(auth()->user()->employee->role->id === 1) 
+                <a class="delete-link" href="/assignTask/delete/{{$task['id']}}"> Borrar </a>
+              @endif
+            @endif
           </td>
         </tr>
         @endforeach
@@ -142,35 +149,35 @@
     
   </div>
   <script>
-  /* Delete Link */
- /*  $(document).ready(function(){
-
-    $('.delete-link').on('click', function (e) {
-      console.log($(this))
+    /* Delete Link */
+    /*  $(document).ready(function(){
+      
+      $('.delete-link').on('click', function (e) {
+        console.log($(this))
         try {
-            e.preventDefault()
-            swal({
-                title: "Borrar",
-                text: "Desea borrar esta entrada?",
-                icon: "warning",
-                buttons: true,
-                dangerMode: true,
-            })
-            .then((willSave) => {
-                try {
-                    if (willSave) {
-                        window.location.href ='/assignTask/delete/{{$task['id']}}'
-                      }
-                } catch (err) {
-                    console.log(err)
-                }
-            });
+          e.preventDefault()
+          swal({
+            title: "Borrar",
+            text: "Desea borrar esta entrada?",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+          })
+          .then((willSave) => {
+            try {
+              if (willSave) {
+                window.location.href ='/assignTask/delete/{{$task['id']}}'
+              }
+            } catch (err) {
+              console.log(err)
+            }
+          });
         } catch (er) {
-            console.log(er)
+          console.log(er)
         }
-    })
-  }) */
-  
+      })
+    }) */
+    
   </script>
   @endsection
